Move this single constants file into a folder where we have others
And now we can delete `src/constants` in favor of `src/lib/constants`
